third_party/blink/web_tests/wpt_internal/css/css-contain/container-queries/container-type-scroll-state-parsing.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Containment Test: Parsing of container-type with scroll-state</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-4/#container-type">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script src="/css/css-contain/container-queries/support/cq-testcommon.js"></script>
<div id="target"></div>
<script>
  setup(() => assert_implements_container_queries());

  test_valid_value('container-type', 'scroll-state');
  test_valid_value('container-type', 'size scroll-state');
  test_valid_value('container-type', 'scroll-state inline-size', 'inline-size scroll-state');

  test_invalid_value('container-type', 'scroll-state scroll-state');
  test_invalid_value('container-type', 'normal scroll-state');
  test_invalid_value('container-type', 'inline-size block-size scroll-state');
</script>
